fix -Fixed a bug in Exercise 1-22.
The function was skipped over every other character because the first  in the while loop already consumed one. This cause a character to be missed, and EOF handling didn't work properly. That was my mistake because I didn't break down the problem correctly.

In programming, we should break a problem into smaller parts and carefully think through each one seperately to ensure that no part of the program or its possible scenarios remains incomplete in the end.